requirements.txt
python-gitlab==4.6.0
requests==2.32.3
